Created example domain for testing purposes.
