+++ apps/event_sourcing/mix.exs
@@ -11,7 +11,8 @@ defmodule EventSourcing.MixProject do
       lockfile: "../../mix.lock",
       elixir: "~> 1.8",
       start_permanent: Mix.env() == :prod,
-      deps: deps()
+      deps: deps(),
+      elixirc_paths: elixirc_paths(Mix.env())
     ]
   end
 
@@ -27,4 +28,7 @@ defmodule EventSourcing.MixProject do
       {:ecto_sql, "~> 3.0"}
     ]
   end
+
+  defp elixirc_paths(:test), do: ["lib", "test/support"]
+  defp elixirc_paths(_), do: ["lib"]
 end
