+++ apps/event_sourcing/test/event_sourcing/aggregates_test.exs
@@ -3,40 +3,12 @@ defmodule EventSourcing.AggregatesTest do
 
   alias EventSourcing.Aggregates
   alias Ecto.UUID
+  alias EventSourcing.Support.Counters.Aggregates.Counter
+  alias EventSourcing.Support.Counters.Commands.Increment
+  alias EventSourcing.Support.Counters.Events.Incremented
+  alias EventSourcing.Support.EventStore.InMemoryEventStore
 
   @registry EventSourcing.AggregateRegistry
-
-  defmodule Increment do
-    defstruct [:uuid, :counter_uuid, :test_pid]
-  end
-
-  defmodule Incremented do
-    defstruct [:uuid, :counter_uuid, :test_pid]
-  end
-
-  defmodule Counter do
-    @behaviour EventSourcing.Aggregate
-    defstruct [:uuid, value: 0]
-
-    def execute(%Counter{} = counter, %Increment{test_pid: pid}) do
-      %Incremented{counter_uuid: counter.uuid, test_pid: pid}
-    end
-
-    def apply(%Counter{value: value} = counter, %Incremented{}) do
-      %{counter | value: value + 1}
-    end
-  end
-
-  defmodule EventStore do
-    @behaviour EventSourcing.EventStore
-
-    def put(uuid, %{test_pid: pid} = event) when is_pid(pid) do
-      send(pid, {:store_put, uuid, event})
-      :ok
-    end
-
-    def get(_uuid), do: []
-  end
 
   describe "execute_command/2" do
     setup :aggregate
@@ -84,7 +56,7 @@ defmodule EventSourcing.AggregatesTest do
       aggregate: {_mod, uuid} = aggregate,
       command: command
     } do
-      {event, _} = Aggregates.execute_command(aggregate, command, store: EventStore)
+      {event, _} = Aggregates.execute_command(aggregate, command, store: InMemoryEventStore)
 
       assert_receive {:store_put, ^uuid, ^event}
     end
