+++ apps/event_sourcing/test/event_sourcing/dispatcher_test.exs
@@ -2,38 +2,10 @@ defmodule EventSourcing.DispatcherTest do
   use ExUnit.Case
 
   alias Ecto.UUID
-
-  defmodule Increment do
-    defstruct [:uuid, :counter_uuid, :test_pid]
-  end
-
-  defmodule Decrement do
-    defstruct [:uuid, :counter_uuid, :test_pid]
-  end
-
-  defmodule Incremented do
-    defstruct [:uuid, :counter_uuid]
-  end
-
-  defmodule Counter do
-    @behaviour EventSourcing.Aggregate
-    defstruct [:uuid, value: 0]
-
-    def execute(%Counter{} = counter, %Increment{test_pid: pid}) do
-      send(pid, {:aggregate_called, __MODULE__})
-      %Incremented{counter_uuid: counter.uuid}
-    end
-
-    def apply(%Counter{value: value} = counter, %Incremented{}) do
-      %{counter | value: value + 1}
-    end
-  end
-
-  defmodule Dispatcher do
-    use EventSourcing.Dispatcher
-
-    dispatch Increment, to: Counter, identity: :counter_uuid
-  end
+  alias EventSourcing.Support.Counters.Aggregates.Counter
+  alias EventSourcing.Support.Counters.Commands.Increment
+  alias EventSourcing.Support.Counters.Commands.Decrement
+  alias EventSourcing.Support.Dispatcher
 
   describe "dispatch/2 macro" do
     test "defines dispatch/1 function" do
