+++ apps/event_sourcing/test/event_sourcing/event_store/agent_event_store_test.exs
@@ -2,10 +2,7 @@ defmodule EventSourcing.EventStore.AgentEventStoreTest do
   use ExUnit.Case
 
   alias EventSourcing.EventStore.AgentEventStore
-
-  defmodule Incremented do
-    defstruct [:uuid, :counter_uuid]
-  end
+  alias EventSourcing.Support.Counters.Events.Incremented
 
   setup do
     start_supervised(AgentEventStore)
