+++ apps/event_sourcing/test/support/counters/aggregates/counter.ex
@@ -1,0 +1,17 @@
+defmodule EventSourcing.Support.Counters.Aggregates.Counter do
+  @behaviour EventSourcing.Aggregate
+
+  alias EventSourcing.Support.Counters.Commands.Increment
+  alias EventSourcing.Support.Counters.Events.Incremented
+
+  defstruct [:uuid, value: 0]
+
+  def execute(%__MODULE__{} = counter, %Increment{test_pid: pid}) do
+    send(pid, {:aggregate_called, __MODULE__})
+    %Incremented{counter_uuid: counter.uuid, test_pid: pid}
+  end
+
+  def apply(%__MODULE__{value: value} = counter, %Incremented{}) do
+    %{counter | value: value + 1}
+  end
+end
