+++ apps/event_sourcing/test/support/counters/commands/decrement.ex
@@ -1,0 +1,3 @@
+defmodule EventSourcing.Support.Counters.Commands.Decrement do
+  defstruct [:uuid, :counter_uuid, :test_pid]
+end
