+++ apps/event_sourcing/test/support/counters/events/incremented.ex
@@ -1,0 +1,3 @@
+defmodule EventSourcing.Support.Counters.Events.Incremented do
+  defstruct [:uuid, :counter_uuid, :test_pid]
+end
