+++ apps/event_sourcing/test/support/event_store/in_memory_event_store.ex
@@ -1,0 +1,10 @@
+defmodule EventSourcing.Support.EventStore.InMemoryEventStore do
+  @behaviour EventSourcing.EventStore
+
+  def put(uuid, %{test_pid: pid} = event) when is_pid(pid) do
+    send(pid, {:store_put, uuid, event})
+    :ok
+  end
+
+  def get(_uuid), do: []
+end
